requirements.txt
msedge_selenium_tools==3.141.4
selenium==3.141.0
